13. Lockers 59

+++ main.h
@@ -44,7 +44,7 @@ int16_t t; // zmienna pomocnicza wykorzystana w pętlach for do iteracji, może 
 int8_t u; //inna (dodatkowa) zmienna pomocnicza
 int8_t w; //inna (dodatkowa) zmienna pomocnicza
 int8_t s;//inna (dodatowa zmienna)
-int8_t c;//inna (dodatowa zmienna)
+int16_t c;//inna (dodatowa zmienna)
 
 uint8_t refresh_screen;
